scripts/requirements.2023-08-13.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
boilerpy3==1.0.6
boto3==1.28.27
botocore==1.31.27
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
camel-converter==3.0.2
canals==0.3.2
catalogue==2.0.9
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
cmake==3.27.2
cmudict==1.0.13
colorama==0.4.6
comm==0.1.4
commonmark==0.9.1
confection==0.1.1
ConfigArgParse==1.7
contourpy==1.1.0
coreferee==1.4.1
crashtest==0.4.1
cryptography==41.0.3
cssutils==2.7.1
cycler==0.11.0
cymem==2.0.7
Cython==3.0.0
cytoolz==0.12.2
dataclasses-json==0.5.14
dataframe-image==0.1.12
datasets==2.14.4
dateparser==1.1.8
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.4
docopt==0.6.2
docutils==0.20.1
dulwich==0.21.5
edit-distance==1.0.6
einops==0.6.1
environment==1.0.0
evaluate==0.4.0
Events==0.5
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.2
farm-haystack==1.19.0
fastjsonschema==2.18.0
filelock==3.12.2
floret==0.10.4
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h5py==3.9.0
html2image==2.0.4.2
html2text==2020.1.16
html5lib==1.1
huggingface-hub==0.16.4
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
importlib-resources==5.13.0
inflect==7.0.0
installer==0.7.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.0
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonlines==3.1.0
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==23.13.1
kiwisolver==1.4.4
langchain==0.0.265
langcodes==3.3.0
langsmith==0.0.22
lazy-imports==0.3.1
lit==16.0.6
llvmlite==0.40.1
lockfile==0.12.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.28.2
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nessvec==0.0.12
nest-asyncio==1.5.7
networkx==3.1
-e git+ssh://git@gitlab.com/tangibleai/nlpia2@d630c81d4eb3f0fd8ed9ce0235216a74fe2101d2#egg=nlpia2
nlpia2-wikipedia==1.5.6
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
num2words==0.5.12
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parsimonious==0.10.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
portalocker==2.7.0
posthog==3.0.1
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prompthub-py==4.0.0
pronouncing==0.2.0
protobuf==4.24.0
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
PyPDF2==3.0.1
pyphen==0.14.0
pyproject-api==1.5.3
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
quantulum3==0.9.0
rank-bm25==0.2.2
rapidfuzz==2.13.7
recommonmark==0.7.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-cache==0.9.8
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
s3transfer==0.6.2
sacrebleu==2.3.1
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
spacy==3.5.4
spacy-experimental==0.6.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.8
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
srsly==2.4.7
sseclient-py==1.7.2
stack-data==0.6.2
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
textacy==0.13.0
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchdata==0.6.1
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.3
tox==4.8.0
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
trove-classifiers==2023.8.7
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
Unidecode==1.3.6
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.16
validators==0.21.2
virtualenv==20.24.3
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==4.0.8
wikipedia==1.4.0
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
